+++ packages/client/src/Call.ts
@@ -72,6 +72,7 @@ import {
 import {
   BehaviorSubject,
   debounce,
+  filter,
   map,
   pairwise,
   Subject,
@@ -371,10 +372,13 @@ export class Call {
    * Leave the call and stop the media streams that were published by the call.
    */
   leave = async ({ reject = false }: CallLeaveOptions = {}) => {
-    // TODO: handle case when leave is called during JOINING
     const callingState = this.state.callingState;
     if (callingState === CallingState.LEFT) {
       throw new Error('Cannot leave call that has already been left.');
+    }
+
+    if (callingState === CallingState.JOINING) {
+      await this.assertCallJoined();
     }
 
     if (this.ringing) {
@@ -804,6 +808,7 @@ export class Call {
     if (!this.publisher) {
       this.publisher = new Publisher({
         sfuClient,
+        dispatcher: this.dispatcher,
         state: this.state,
         connectionConfig,
         isDtxEnabled,
@@ -848,6 +853,7 @@ export class Call {
             subscriberSdp: sdp || '',
             clientDetails: getClientDetails(),
             migration,
+            fastReconnect: false,
           });
         });
 
@@ -925,14 +931,6 @@ export class Call {
         unsubscribe();
         reject(new Error('Waiting for "joinResponse" has timed out'));
       }, timeout);
-    });
-  };
-
-  private assertCallJoined = () => {
-    return new Promise<void>((resolve) => {
-      this.state.callingState$
-        .pipe(takeWhile((state) => state !== CallingState.JOINED, true))
-        .subscribe(() => resolve());
     });
   };
 
@@ -1219,6 +1217,17 @@ export class Call {
    */
   updatePublishQuality = async (enabledRids: string[]) => {
     return this.publisher?.updateVideoPublishQuality(enabledRids);
+  };
+
+  private assertCallJoined = () => {
+    return new Promise<void>((resolve) => {
+      this.state.callingState$
+        .pipe(
+          takeWhile((state) => state !== CallingState.JOINED, true),
+          filter((s) => s === CallingState.JOINED),
+        )
+        .subscribe(() => resolve());
+    });
   };
 
   /**