+++ packages/react-native-sdk/src/providers/MediaStreamManagement/index.tsx
@@ -96,16 +96,32 @@ export const MediaStreamManagement = ({ children }: PropsWithChildren<{}>) => {
   const [isCameraOnFrontFacingMode, setIsCameraOnFrontFacingMode] =
     useState(true);
 
-  const [initAudioEnabled, setInitialAudioEnabled] = useState<boolean>(
-    isMicPermissionGranted$.getValue() &&
-      !!call?.permissionsContext?.hasPermission(OwnCapability.SEND_AUDIO),
-  );
+  const [initAudioEnabled, setInitialAudioEnabled] = useState<boolean>(() => {
+    const hasNativePermission = isMicPermissionGranted$.getValue();
+    const hasUserPermission = !!call?.permissionsContext?.hasPermission(
+      OwnCapability.SEND_AUDIO,
+    );
+    const metaDataSettings = call?.data?.settings?.audio.mic_default_on;
+    if (metaDataSettings !== undefined) {
+      return hasNativePermission && hasUserPermission && metaDataSettings;
+    }
+    return hasNativePermission && hasUserPermission;
+  });
 
-  const [initVideoEnabled, setInitialVideoEnabled] = useState<boolean>(
-    isCameraPermissionGranted$.getValue() &&
-      call?.type !== 'audio_room' &&
-      !!call?.permissionsContext?.hasPermission(OwnCapability.SEND_VIDEO),
-  );
+  const [initVideoEnabled, setInitialVideoEnabled] = useState<boolean>(() => {
+    if (call?.type === 'audio_room') {
+      return false;
+    }
+    const hasNativePermission = isCameraPermissionGranted$.getValue();
+    const hasUserPermission = !!call?.permissionsContext?.hasPermission(
+      OwnCapability.SEND_VIDEO,
+    );
+    const metaDataSettings = call?.data?.settings?.video.camera_default_on;
+    if (metaDataSettings !== undefined) {
+      return hasNativePermission && hasUserPermission && metaDataSettings;
+    }
+    return hasNativePermission && hasUserPermission;
+  });
 
   const publishVideoStream = useVideoPublisher({
     initialVideoMuted: !initVideoEnabled,