+++ sample-apps/react-native/dogfood/src/screens/AudioRoom/RoomList.tsx
@@ -38,7 +38,7 @@ const RoomList = (props: Props) => {
   const nextPage = useRef<string>();
 
   const queryLiveCalls = useCallback(async () => {
-    if (!client || client.user?.id === undefined) {
+    if (!client) {
       return;
     }
     setLoadingCalls(true);